PlanetConverter renamed to SPlanetConverter
